PR suggestions

+++ tests/util.py
@@ -51,7 +51,7 @@ def download_model(url: str, name: str | None = None) -> str:
     return path
 
 
-def extract_zip(path: str, rel_model_dir: Path, name: str):
+def extract_zip(path: str, rel_model_path: Path | str, name: str):
     if not zipfile.is_zipfile(path):
         print(f"Skipping {path} because it is not a zip file.")
         return
@@ -63,7 +63,7 @@ def extract_zip(path: str, rel_model_dir: Path, name: str):
     with zipfile.ZipFile(path, "r") as zip_ref:
         with tempfile.TemporaryDirectory() as tmpdirname:
             zip_ref.extractall(tmpdirname)
-            model_path = Path(tmpdirname) / rel_model_dir
+            model_path = Path(tmpdirname) / rel_model_path
             assert model_path.exists(), f"Expected {model_path} to exist."
             model_path.rename(MODEL_DIR / name)
             return model_path
@@ -95,14 +95,13 @@ class ModelFile:
         return ModelFile(name).download(url)
 
     @staticmethod
-    def from_url_zip(
-        url: str, name: str | None = None, rel_model_dir: Path | None = None
-    ):
+    def from_url_zip(url: str, rel_model_path: Path | str, name: str | None = None):
+        name = os.path.basename(rel_model_path) if name is None else name
+        if (MODEL_DIR / name).exists():
+            return ModelFile(name)
         path = download_model(url, "temp.zip")
         print(f"Extracting {path}...")
-        extract_zip(
-            path, rel_model_dir or Path(name or ""), name or get_url_file_name(url)
-        )
+        extract_zip(path, rel_model_path or name, name)
         os.remove(path)
         return ModelFile(name or get_url_file_name(url))
